+++ src/components/tests/app.test.js
@@ -1,13 +1,8 @@
 import React from "react";
-import Enzyme, {shallow} from "enzyme";
-import Adapter from "enzyme-adapter-react-16";
-import Films from "./Films";
+import renderer from "react-test-renderer";
+import App from '../App';
 
-Enzyme.configure({
-  adapter: new Adapter(),
-});
-
-const filmsData =
+const films = [
   {
     title: `Some film2`,
     year: `2014`,
@@ -39,51 +34,17 @@ const filmsData =
         },
       ]
     },
-  };
+  },
+];
 
-describe(`Films`, () => {
-  it(`Should title be pressed`, () => {
-    const index = 1;
-    const onFilmsTitleClick = jest.fn();
-    const onMouseCard = jest.fn();
-
-    const filmsElement = shallow(
-        <Films
-          film={filmsData}
-          key={index}
-          onFilmsTitleClick={onFilmsTitleClick}
-          onMouseCard={onMouseCard}
+describe(`Foo`, () => {
+  it(`<App /> should render main page in app`, () => {
+    const tree = renderer.create(
+        <App
+          filmsData={films}
         />
-    );
+    ).toJSON();
 
-    const filmsTitle = filmsElement.find(`a.small-movie-card__link`);
-
-    filmsTitle.forEach((filmTitle) => {
-      filmTitle.simulate(`click`);
-    });
-  });
-  it(`Should card be hover`, () => {
-    const onFilmsTitleClick = jest.fn();
-    const onMouseCard = jest.fn();
-    const index = 1;
-
-    const films = shallow(
-        <Films
-          film={filmsData}
-          key={index}
-          onFilmsTitleClick={onFilmsTitleClick}
-          onMouseCard={onMouseCard}
-        />
-    );
-    const filmsArticles = films.find(`article.small-movie-card.catalog__movies-card`);
-    filmsArticles.forEach((filmsArticle) => {
-      expect(films.state(`onMouseCard`)).toBe(false);
-      filmsArticle.simulate(`mouseover`);
-
-      expect(films.state(`onMouseCard`)).toBe(true);
-      filmsArticle.simulate(`mouseleave`);
-      expect(films.state(`onMouseCard`)).toBe(false);
-    });
+    expect(tree).toMatchSnapshot();
   });
 });
-
